requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
alphashape==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
BareNecessities==0.2.8
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
certifi==2024.12.14
cffi==1.17.0
charset-normalizer==3.4.0
click==8.1.7
click-log==0.4.0
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cryptography==44.0.0
cryptography-fernet==0.1.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.7
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.15.4
fiona==1.10.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
folium==0.19.5
fonttools==4.47.2
frozenlist==1.5.0
geographiclib==2.0
geoip2==4.8.1
geopandas==1.0.1
geopy==2.4.1
gotrue==2.11.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hupper==1.12.1
hyperframe==6.0.1
idna==3.10
ipykernel==6.29.4
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
Js2Py==0.74
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
kneed==0.8.5
Mail==2.1.0
Mako==1.3.8
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.8.2
matplotlib-inline==0.1.7
maxminddb==2.6.3
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
munkres==1.1.4
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nbimporter==0.3.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.3
openpyxl==3.1.2
osmnx==2.0.1
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.2.0
pipdeptree==2.24.0
pipreqs==0.5.0
pipwin==0.5.2
platformdirs==4.2.1
postgrest==0.18.0
prompt-toolkit==3.0.43
propcache==0.2.1
psutil==5.9.8
psycopg2==2.9.10
pure-eval==0.2.2
py-search==2.8.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydotplus==2.0.2
Pygments==2.18.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyjsparser==2.7.1
PyJWT==2.9.0
PyMySQL==1.1.1
pyogrio==0.10.0
pyparsing==3.1.1
PyPrind==2.11.3
pyproj==3.7.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pySmartDL==1.3.4
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
pyzmq==26.0.3
realtime==2.0.6
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
rpy2==3.5.16
rtree==1.4.0
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
shapely==2.0.7
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.4
traitlets==5.14.3
trimesh==4.6.6
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
virtualenv==20.26.3
waitress==3.0.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
Werkzeug==3.1.3
xlwt==1.3.0
xyzservices==2025.1.0
yarg==0.1.9
yarl==1.18.3
